backend/pyproject.toml
[project]
name = "ai-reddit-backend"
version = "1.0.0"
description = "AI-Native Reddit MVP Backend"
requires-python = ">=3.8"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn[standard]==0.24.0",
    "sqlalchemy==2.0.23",
    "pydantic==2.5.0",
    "python-multipart==0.0.6",
    "anthropic==0.7.8",
    "numpy>=1.21.0",
    "httpx==0.25.2",
    "python-dotenv==1.0.0",
    "faker==20.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
]

[tool.uv]
dev-dependencies = [
    "pytest==7.4.3",
    "pytest-asyncio==0.21.1",
]
